Beginings of code to track clients.
In the system_dict I have no idea what RHN uses the uuid field for.  Seems
to always be empty.  The rhnuuid field is the UUID created when
up2date is first installed (%post of the package).  

RHN uses some strange system ID number in the syste_id field of the systemid
structure.  I do not want to use internal database numbers there so
we use the clients UUID.  Need to correct the SysID code to handle those
strings.



git-svn-id: http://current.tigris.org/svn/current/trunk@220 7112b371-5efc-0310-9db1-efd4ebc68ee9