From Pauline:
Attached a patch which adds NVRE to PACKAGE, and only a package_id
to INSTALLED.

You will note a bit of code duplication regarding currentdb._getPackageId()
and currentdb._insertPackage() vs profiledb/_xxx(). I'm not sure how
to handle this neatly. Except by making them public in currentdb
and using them from profile, but that would entail getting an object
for it etc etc. Not fun. Enfin, just a proof of concept, you as
the architect may make the diffcult decisions :)


git-svn-id: http://current.tigris.org/svn/current/trunk@266 7112b371-5efc-0310-9db1-efd4ebc68ee9