From Pauline:
The != 1 check is tricky. I my case I defined an extra channel for my
IEL4 distro, and of course with the same os_release. Problem is that the
subscribed client suddely has 2 matching channels and is punished for
it :(

The other thing I ran into is that cannon(athlon) == i386 according to
me. Tell me if I'm wrong, but at least my AMD machine can register now...


git-svn-id: http://current.tigris.org/svn/current/trunk@268 7112b371-5efc-0310-9db1-efd4ebc68ee9